+++ src/AckApi.ts
@@ -1,13 +1,19 @@
 import 'rxjs/add/operator/toPromise';
 import { Injectable, EventEmitter } from '@angular/core';
-import { HttpClient, HttpRequest, HttpResponse, HttpEvent } from '@angular/common/http'
+
+import { Http, Response, Request } from '@angular/http';
+//Angular5
+//import { HttpClient, HttpRequest, HttpResponse, HttpEvent } from '@angular/common/http'
 
 import { AckCache } from './AckCache';
 import { AckQue } from './AckQue';
 
 /** Http util with offline config for request failures */
 @Injectable() export class AckApi{
-  response:EventEmitter<HttpResponse<HttpEvent<Event>>> = new EventEmitter()
+  //Angular5
+  //response:EventEmitter<HttpResponse<HttpEvent<Event>>> = new EventEmitter()
+
+  response:EventEmitter<Response> = new EventEmitter()
   AuthError:EventEmitter<Error> = new EventEmitter()
   ApiError:EventEmitter<Error> = new EventEmitter()
   AckCache
@@ -28,7 +34,7 @@ import { AckQue } from './AckQue';
     }
   }
 
-  constructor(public HttpClient:HttpClient){
+  constructor(public HttpClient:Http){
     this.AckCache = new AckCache()
     this.AckQue = new AckQue()
     this.paramConfig()
@@ -184,15 +190,17 @@ import { AckQue } from './AckQue';
       headers:{}//when sending a file 'Content-Type':undefined so that no content-type header is sent
     }
   */
-  _fetch(cfg):Promise<HttpResponse<HttpEvent<Event>>> {
+  //Angular 5 : Promise<HttpResponse<HttpEvent<Event>>>
+  _fetch(cfg):Promise<Response> {
     upgradeConfig(cfg)
 
-    const request = new HttpRequest(
+    const request = new Request(cfg)
+    /*const request = new HttpRequest(
       cfg.method,
       cfg.url,
       cfg.body,
       cfg
-    )//cfg
+    )*/
  
     return this.HttpClient.request( request )
     .toPromise()